requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.2.25
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
faiss-cpu==1.13.2
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
minio==7.2.20
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pycparser==3.0
pycryptodome==3.23.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.17.1
six==1.17.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
